requirements.txt
requests>=2.31.0
base58>=2.1.1
solders>=0.18.0
google-genai>=1.0.0
websockets>=12.0
python-dotenv>=1.0.0